// src/generated/castle_contracts.ts
export interface Shortcut {
  category: string;
  label: string;
  description: string;
  href: string;
}